# Optimize disk.FileReader