This makes larger pebble reads ~5% faster.  There are 3 optimizations:
1. Skip a Stat call, since it's not necessary. `io.SectionedReader` works with `MaxInt64` as the length.
2. Stop tracing every `Read()` call. It doesn't seem that useful and it uses 0.3% of app CPU. It produces traces like this: 
![image](https://github.com/user-attachments/assets/28413dbc-33ba-431f-8bb5-f8d3702482ad)
 and I suspect that tracing takes a significant portion of time for each read. If we still think this tracing is useful, we could instead trace the whole lifespan of the reader, but I think that's better done at a higher level.
3. Skipping tracing allows us to directly return an os.File when offset and length are 0. This skips one layer of indirection, and allows the returned reader to also implement `WriteTo`.

Benchmark results from `//enterprise/server/test/performance/cache:cache_test`:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor            
                            │  cachebase  │            cachesimple             │
                            │   sec/op    │   sec/op     vs base               │
Get/Pebble10-24               6.131µ ± 1%   6.093µ ± 2%       ~ (p=0.317 n=13)
Get/Pebble100-24              6.545µ ± 2%   6.456µ ± 2%  -1.36% (p=0.012 n=13)
Get/Pebble1000-24             7.656µ ± 1%   7.541µ ± 1%  -1.50% (p=0.003 n=13)
Get/Pebble10000-24            25.32µ ± 1%   23.67µ ± 1%  -6.50% (p=0.000 n=13)
Get/DistPebble10-24           92.78µ ± 1%   92.86µ ± 1%       ~ (p=0.362 n=13)
Get/DistPebble100-24          93.67µ ± 1%   94.17µ ± 1%       ~ (p=1.000 n=13)
Get/DistPebble1000-24         97.20µ ± 2%   96.78µ ± 1%       ~ (p=0.362 n=13)
Get/DistPebble10000-24        133.9µ ± 1%   126.7µ ± 1%  -5.38% (p=0.000 n=13)
GetMulti/Pebble10-24          390.0µ ± 2%   387.2µ ± 2%       ~ (p=0.101 n=13)
GetMulti/Pebble100-24         430.7µ ± 2%   427.9µ ± 3%       ~ (p=0.139 n=13)
GetMulti/Pebble1000-24        530.7µ ± 2%   533.2µ ± 2%       ~ (p=0.624 n=13)
GetMulti/Pebble10000-24       2.289m ± 1%   2.131m ± 1%  -6.89% (p=0.000 n=13)
GetMulti/DistPebble10-24      495.7µ ± 6%   469.5µ ± 8%       ~ (p=0.223 n=13)
GetMulti/DistPebble100-24     711.0µ ± 3%   709.3µ ± 3%       ~ (p=0.243 n=13)
GetMulti/DistPebble1000-24    931.9µ ± 2%   905.5µ ± 3%  -2.84% (p=0.002 n=13)
GetMulti/DistPebble10000-24   3.234m ± 2%   3.075m ± 1%  -4.90% (p=0.000 n=13)
geomean                       160.3µ        156.6µ       -2.30%

                            │  cachebase   │             cachesimple             │
                            │     B/s      │     B/s       vs base               │
Get/Pebble10-24               1.554Mi ± 1%   1.564Mi ± 2%       ~ (p=0.366 n=13)
Get/Pebble100-24              14.57Mi ± 2%   14.77Mi ± 2%  +1.37% (p=0.011 n=13)
Get/Pebble1000-24             124.6Mi ± 1%   126.5Mi ± 1%  +1.54% (p=0.003 n=13)
Get/Pebble10000-24            376.7Mi ± 1%   402.9Mi ± 1%  +6.95% (p=0.000 n=13)
Get/DistPebble10-24           107.4Ki ± 0%   107.4Ki ± 0%       ~ (p=1.000 n=13)
Get/DistPebble100-24          1.020Mi ± 2%   1.011Mi ± 1%       ~ (p=0.891 n=13)
Get/DistPebble1000-24         9.813Mi ± 2%   9.851Mi ± 1%       ~ (p=0.328 n=13)
Get/DistPebble10000-24        71.22Mi ± 1%   75.27Mi ± 1%  +5.69% (p=0.000 n=13)
GetMulti/Pebble10-24          2.441Mi ± 2%   2.460Mi ± 2%       ~ (p=0.148 n=13)
GetMulti/Pebble100-24         22.14Mi ± 2%   22.29Mi ± 3%       ~ (p=0.142 n=13)
GetMulti/Pebble1000-24        179.7Mi ± 2%   178.9Mi ± 2%       ~ (p=0.624 n=13)
GetMulti/Pebble10000-24       416.6Mi ± 1%   447.4Mi ± 1%  +7.41% (p=0.000 n=13)
GetMulti/DistPebble10-24      1.926Mi ± 6%   2.031Mi ± 7%       ~ (p=0.190 n=13)
GetMulti/DistPebble100-24     13.42Mi ± 3%   13.45Mi ± 3%       ~ (p=0.248 n=13)
GetMulti/DistPebble1000-24    102.3Mi ± 3%   105.3Mi ± 3%  +2.93% (p=0.002 n=13)
GetMulti/DistPebble10000-24   294.9Mi ± 2%   310.1Mi ± 1%  +5.16% (p=0.000 n=13)
geomean                       18.84Mi        19.27Mi       +2.32%

                            │  cachebase   │             cachesimple              │
                            │     B/op     │     B/op      vs base                │
Get/Pebble10-24               4.473Ki ± 0%   4.473Ki ± 0%        ~ (p=1.000 n=13)
Get/Pebble100-24              4.612Ki ± 0%   4.612Ki ± 0%        ~ (p=0.900 n=13)
Get/Pebble1000-24             5.816Ki ± 0%   5.815Ki ± 0%        ~ (p=0.653 n=13)
Get/Pebble10000-24            16.99Ki ± 0%   14.81Ki ± 0%  -12.82% (p=0.000 n=13)
Get/DistPebble10-24           22.37Ki ± 0%   22.37Ki ± 0%        ~ (p=0.342 n=13)
Get/DistPebble100-24          22.84Ki ± 0%   22.83Ki ± 0%        ~ (p=0.237 n=13)
Get/DistPebble1000-24         26.75Ki ± 0%   26.75Ki ± 0%        ~ (p=0.830 n=13)
Get/DistPebble10000-24        45.34Ki ± 0%   43.12Ki ± 0%   -4.90% (p=0.000 n=13)
GetMulti/Pebble10-24          364.2Ki ± 0%   364.2Ki ± 0%        ~ (p=0.830 n=13)
GetMulti/Pebble100-24         378.8Ki ± 0%   378.9Ki ± 0%        ~ (p=0.317 n=13)
GetMulti/Pebble1000-24        499.2Ki ± 0%   499.2Ki ± 0%        ~ (p=1.000 n=13)
GetMulti/Pebble10000-24       1.579Mi ± 0%   1.366Mi ± 0%  -13.45% (p=0.000 n=13)
GetMulti/DistPebble10-24      390.8Ki ± 7%   367.9Ki ± 9%        ~ (p=0.186 n=13)
GetMulti/DistPebble100-24     553.8Ki ± 0%   553.7Ki ± 0%        ~ (p=0.762 n=13)
GetMulti/DistPebble1000-24    826.9Ki ± 1%   827.2Ki ± 1%        ~ (p=0.724 n=13)
GetMulti/DistPebble10000-24   2.860Mi ± 1%   2.651Mi ± 1%   -7.32% (p=0.000 n=13)
geomean                       98.11Ki        95.28Ki        -2.89%

                            │  cachebase   │              cachesimple               │
                            │  allocs/op   │  allocs/op    vs base                  │
Get/Pebble10-24                 90.00 ± 0%    90.00 ±  0%        ~ (p=1.000 n=13) ¹
Get/Pebble100-24                90.00 ± 0%    90.00 ±  0%        ~ (p=1.000 n=13) ¹
Get/Pebble1000-24               90.00 ± 0%    90.00 ±  0%        ~ (p=1.000 n=13) ¹
Get/Pebble10000-24             118.00 ± 0%    91.00 ±  0%  -22.88% (p=0.000 n=13)
Get/DistPebble10-24             425.0 ± 0%    425.0 ±  0%        ~ (p=1.000 n=13) ¹
Get/DistPebble100-24            425.0 ± 0%    425.0 ±  0%        ~ (p=1.000 n=13) ¹
Get/DistPebble1000-24           425.0 ± 0%    425.0 ±  0%        ~ (p=1.000 n=13) ¹
Get/DistPebble10000-24          448.0 ± 0%    421.0 ±  0%   -6.03% (p=0.000 n=13)
GetMulti/Pebble10-24           7.113k ± 0%   7.113k ±  0%        ~ (p=1.000 n=13)
GetMulti/Pebble100-24          7.113k ± 0%   7.113k ±  0%        ~ (p=0.220 n=13)
GetMulti/Pebble1000-24         7.114k ± 0%   7.114k ±  0%        ~ (p=1.000 n=13) ¹
GetMulti/Pebble10000-24        9.919k ± 0%   7.219k ±  0%  -27.22% (p=0.000 n=13)
GetMulti/DistPebble10-24       6.591k ± 8%   6.156k ± 10%        ~ (p=0.122 n=13)
GetMulti/DistPebble100-24      9.217k ± 0%   9.217k ±  0%        ~ (p=1.000 n=13)
GetMulti/DistPebble1000-24     9.227k ± 0%   9.227k ±  0%        ~ (p=0.633 n=13)
GetMulti/DistPebble10000-24   12.114k ± 0%   9.425k ±  0%  -22.20% (p=0.000 n=13)
geomean                        1.306k        1.230k         -5.82%
```